+++ backend/src/main/java/com/chamchi/backend/repository/category/CategoryCountRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 
-public interface CategoryCountRepository extends JpaRepository<CategoryCount, Long> {
+public interface CategoryCountRepository extends JpaRepository<CategoryCount, Integer> {
     @Query(value = "SELECT c.id , c.name ,count(c.id) AS count FROM Category c INNER JOIN Post p ON c.id = p.category.id  WHERE p.isDeleted = 'N' GROUP BY c.id")
     Optional<List<Object[]>> findAllCategoryWithCount();
 }
